+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - Fix: find definition will find registration of unnamespaced keyword.
   - Fix to update unused-public-var lint on registered keywords as usages change in other files. #1018
   - Fix to navigate to var defined by declare, when there aren't any later defs. #1107
+  - Fix to always go to the definition of the correct var imported by potemkin. #1020
 
 ## 2022.06.29-19.32.13